Fix uses of `long` type in the code (for compatibility with MSVS)

+++ Common/libsrc/aastr-0.1.1/aarot.c
@@ -41,9 +41,9 @@ _aa_rotate_bitmap (BITMAP *_src, BITMAP *_dst, int _x, int _y, fixed _angle,
   int rscinc, rscdd, rsci1, rsci2;
   int sxinc, sxdd, sxi1, sxi2;
   int syinc, sydd, syi1, syi2;
-  unsigned long num;
-  void (*add) (BITMAP *_src, int _sx1, int _sx2, int _sy1, int _sy2, unsigned long _num);
-  void (*put) (unsigned long _addr, int _x);
+  uint32_t num;
+  void (*add) (BITMAP *_src, int _sx1, int _sx2, int _sy1, int _sy2, uint32_t _num);
+  void (*put) (uintptr_t _addr, int _x);
 
   if (_dst->clip)
     {
@@ -477,7 +477,7 @@ _aa_rotate_bitmap (BITMAP *_src, BITMAP *_dst, int _x, int _y, fixed _angle,
   /* Stretch lines.  */
   while (dy < yend)
     {
-      unsigned long daddr = bmp_write_line (_dst, dy);
+      uintptr_t daddr = bmp_write_line (_dst, dy);
 
       if ((ldx < xend) && (rdx >= xbeg))
 	{
